Reverted files' renaming.
